Exported `Promise#then0` for use by babybird-aware downstream libraries.
In theory, a native Promise implementation could use compile-time tricks
to transform `Promise#then` calls to `Promise#then0` calls whenever it
noticed that the result of the call was not used.  But in practice,
clients have to manually indicate this.  Using `then0` instead of `then`
can yield significant performance improvements.